requirements.txt
phidata
python-dotenv
duckduckgo-search
yfinance
pandas
packaging
uvicorn
fastapi
groq
pillow
streamlit

